feat: end of mulligan event (#23)
It turns out as the functionality of mulligan-change that initialized
the number of cards in each deck was extraneous, as that was already
handled by the zone-change line parser. This did not affect our tests
for 2 reasons.

For the test that uses dummy.log, we were using an old format of the
hearthstone log, where the line for mulligan-start came after all the
cards were added to the deck and thus triggered the zone-change line
parser, which resulted in the mulligan-start setting the deck count from
30 to 30.

For the test that uses the newer togwaggle_deck_swap.log, where the cards were
added to the deck after the mulligan, there was a single space at the
end of the line for the mulligan start in the log, which made our parser
ignore it, but the cards were still added through our zone-change line
parser.

Thus, it would make sense to remove the deck count setting of the
MulliganStart parser, keep track of the mulligan start in the
GameTagChange parser, and then use that to set the state of
mulliganActive.